chatbot-evaluation/requirements.txt
# requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
pydantic
pyyaml
sentence-transformers